refactor(components): 비디오 프리뷰 컴포넌트 구조 개선
- VideoSummary → video-preview 디렉토리로 통합
- VideoPreviewList, AnalysisPanel 이동
- Thumbnail 컴포넌트 common으로 이동
- 컴포넌트 재사용성 향상

Related to #71
