+++ src/components/common/video-preview/VideoPreviewList.tsx
@@ -1,16 +1,16 @@
 'use client';
 
-import VideoItem from './VideoItem';
+import VideoPreview from './VideoPreview';
 import {useAuth} from "@/contexts/AuthContext";
-import {VideoSummaryItem as VideoSummaryItemType} from "@/types/video-summary.types";
+import {VideoSummaryItem} from "@/types";
 import LoginCallout from "@components/common/LoginCallout";
 
 interface Props {
     type?: string;
-    data: VideoSummaryItemType[];
+    data: VideoSummaryItem[];
 }
 
-export default function VideoList({type = "trending", data}: Props) {
+export default function VideoPreviewList({type = "trending", data}: Props) {
     const {isLoggedIn} = useAuth();
 
     if (type === "scraps" && !isLoggedIn) {
@@ -33,7 +33,7 @@ export default function VideoList({type = "trending", data}: Props) {
     return (
         <div className="flex flex-col gap-5">
             {data.map((item, i) => (
-                <VideoItem key={item.video?.id ?? i} rank={i + 1} data={item}/>
+                <VideoPreview key={item.video?.id ?? i} rank={i + 1} data={item}/>
             ))}
         </div>
     );
